预测的数字类型返回值修改为Double类型

+++ mulanbay-ai/src/test/java/cn/mulanbay/ai/ml/TestBudgetConsumeMEvaluateProcessor.java
@@ -23,9 +23,9 @@ public class TestBudgetConsumeMEvaluateProcessor {
     @Autowired
     BudgetConsumeMEvaluateProcessor evaluateProcessor;
 
-    private float min = 0f ;
+    private double min = 0f ;
 
-    private float max = 1.5f;
+    private double max = 1.5f;
 
     @Before
     public void init(){
@@ -39,12 +39,12 @@ public class TestBudgetConsumeMEvaluateProcessor {
     //@Test
     @Deprecated
     public void testEvaluate(){
-        float evMin = max;
-        float evMax = min;
+        double evMin = max;
+        double evMax = min;
         for(int month = 1;month<=12;month++){
             for(int score = 0;score<=100;score++){
                 for(int dayIndex = 1;dayIndex<=31;dayIndex++){
-                    Float v = evaluateProcessor.evaluate(month,score,dayIndex);
+                    Double v = evaluateProcessor.evaluate(month,score,dayIndex);
                     String vs = "month="+month+",score="+score+",dayIndex="+dayIndex+",预测值="+v;
                     Assert.notNull(v,"预测值为空,"+vs);
                     Assert.isTrue(v>=min,"预测值低于最低值,"+vs);
@@ -70,9 +70,9 @@ public class TestBudgetConsumeMEvaluateProcessor {
         for(int month = 1;month<=12;month++){
             for(int score = 0;score<=100;score++){
                 for(int dayIndex = 1;dayIndex<=31;dayIndex++){
-                    Map<String,Float> es = evaluateProcessor.evaluateMulti(month,score,dayIndex);
-                    Float predictRate1 = es.get("rate1");
-                    Float predictRate2 = es.get("rate2");
+                    Map<String,Double> es = evaluateProcessor.evaluateMulti(month,score,dayIndex);
+                    Double predictRate1 = es.get("rate1");
+                    Double predictRate2 = es.get("rate2");
                     String vs1 = "month="+month+",score="+score+",dayIndex="+dayIndex+",rate1预测值="+predictRate1;
                     String vs2 = "month="+month+",score="+score+",dayIndex="+dayIndex+",rate2预测值="+predictRate2;
                     Assert.notNull(predictRate1,"预测值为空,"+vs1);
